testing/mochitest/tests/Harness_sanity/test_sanity_manifest.html



<!DOCTYPE HTML>
<html>
<head>
  <title>SimpleTest.expected = 'fail' test</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<script class="testbody" type="text/javascript">
ok(false, "We expect this to fail");

</script>
</body>
</html>
